requirements-dev.txt
# Development Requirements (for training/eval on laptop/workstation)
# Install with: pip install -r requirements-dev.txt

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Metrics and Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
jupyter>=1.0.0

# Optional - for visualization
ipywidgets>=8.0.0
